LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/safe-passing-of-structured-data/shared-array-buffers/window-iframe-messagechannel.https-expected.txt

CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/html/infrastructure/safe-passing-of-structured-data/shared-array-buffers/resources/iframe-messagechannel-site-failure.html because it violates the resource's Cross-Origin-Resource-Policy response header.

Harness Error (TIMEOUT), message = null

PASS postMessaging to a same-origin iframe via MessageChannel allows them to see each others' modifications
TIMEOUT postMessaging to a same-site iframe via MessageChannel should fail Test timed out
NOTRUN postMessaging to a cross-site iframe via MessageChannel should fail
NOTRUN postMessaging with a MessageChannel that's been cross-site should succeed
